styles: fixed by cpplint

+++ source/include/Utils/File.hpp
@@ -10,17 +10,16 @@ concept ByteValueType = std::is_integral_v<ValueType> && sizeof(ValueType) == 1;
 
 template <typename ContainerType>
 concept AppendableBytesContainer =
-    requires(ContainerType a)
-{
-    requires std::ranges::contiguous_range<ContainerType>;
-    requires ByteValueType<typename ContainerType::value_type>;
-    requires std::is_constructible_v<ContainerType>;
-    requires std::is_constructible_v<ContainerType, size_t,
-typename ContainerType::value_type>; // std::string(count, ch),
-// std::vector(count, value)
-a.insert(a.end(), a.begin(), a.begin() + (size_t)1);
-a.resize(a.size());
-};
+    requires(ContainerType a) {
+        requires std::ranges::contiguous_range<ContainerType>;
+        requires ByteValueType<typename ContainerType::value_type>;
+        requires std::is_constructible_v<ContainerType>;
+        requires std::is_constructible_v<ContainerType, size_t,
+                                         typename ContainerType::value_type>; // std::string(count, ch),
+        // std::vector(count, value)
+        a.insert(a.end(), a.begin(), a.begin() + 1);
+        a.resize(a.size());
+    };
 
 template <AppendableBytesContainer ContainerType>
 ContainerType read_file(const std::filesystem::path& path)
